requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.23.4
fake-useragent==1.4.0
python-dateutil==2.8.2
